Git Hook and Ruby SourceStats
Adds a pre-commit git hook file I've begun developing for Speroc commits. This file currently serves to automatically generate some data for the readme at every commit as I've frequently shown myself to be incapable of doing so. It is mostly written in Ruby, although I've decided to group the "assemble_readme" file in the Git folder as it is directly tied to the hook operation.
Also adds in a simple script for determining a basic consideration of the sloc for C/C++ and Ruby files. This script is used within the git hook operation, but is sufficiently divorced that I've decided to split it into a separate folder